admin user and group protected.

+++ application/views/admin/group/index.php
@@ -4,8 +4,6 @@
             <h1>Groups <a href="/admin/group/create" class="btn btn-primary pull-right">Create Group</a></h1>
         </div>
     </div>
-</div>
-<div class="row">
     <div class="col-md-12">
         <?=$this->flash->display()?>
         <table class="table">
@@ -21,8 +19,12 @@
                 <tr>
                     <td><?=$group->name?></td>
                     <td><?=$group->description?></td>
+                <?php if($group->name != $this->ion_auth->get_config('admin_group')): ?>
                     <td><a href="/admin/group/edit/<?=$group->id?>">Edit</a></td>
                     <td><a href="/admin/group/delete/<?=$group->id?>" onclick="return confirmDelete();">Delete</a></td>
+                <?php else: ?>
+                    <td colspan="2">&nbsp;</td>
+                <?php endif; ?>
                 </tr>
             <?php endforeach ?>
             </tbody>
