+++ application/views/admin/user/index.php
@@ -4,8 +4,6 @@
             <h1>Users <a href="/admin/user/create" class="btn btn-primary pull-right">Create User</a></h1>
         </div>
     </div>
-</div>
-<div class="row">
     <div class="col-md-12">
         <?=$this->flash->display()?>
         <table class="table">
@@ -16,6 +14,7 @@
                     <th>Company</th>
                     <th>Email</th>
                     <th>Group</th>
+                    <th>Active</th>
                     <th colspan="2">&nbsp;</th>
                 </tr>
             </thead>
@@ -27,8 +26,13 @@
                     <td><?=$user->company?></td>
                     <td><?=$user->email?></td>
                     <td><?=$user->group_description?></td>
-                    <td><a href="/admin/user/edit/<?=$user->id?>">Edit</a></td>
-                    <td><a href="/admin/user/delete/<?=$user->id?>" onclick="return confirmDelete();">Delete</a></td>
+                    <td><?=($user->active) ? '<i class="icon icon-ok"></i>' : '<i class="icon icon-remove"></i>' ?></td>
+                    <?php if($user->group != $this->ion_auth->get_config('admin_group')): ?>
+                        <td><a href="/admin/user/edit/<?=$user->id?>">Edit</a></td>
+                        <td><a href="/admin/user/delete/<?=$user->id?>" onclick="return confirmDelete();">Delete</a></td>
+                    <?php else: ?>
+                        <td colspan="2">&nbsp;</td>
+                    <?php endif; ?>
                 </tr>
             <?php endforeach ?>
             </tbody>
